support files larger than 2.1GB

+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/media/MediaService.kt
@@ -211,7 +211,7 @@ class MediaServiceImpl(
 
     override suspend fun prepareUploadMedia(content: ByteArrayFlow, contentType: ContentType?): String {
         return "$UPLOAD_MEDIA_CACHE_URI_PREFIX${SecureRandom.nextString(22)}".also { cacheUri ->
-            var fileSize = 0
+            var fileSize = 0L
             mediaStore.addMedia(cacheUri, content.onEach { fileSize += it.size })
             mediaCacheMappingStore.saveMediaCacheMapping(
                 cacheUri,
@@ -237,7 +237,7 @@ class MediaServiceImpl(
             width = thumbnail.width,
             height = thumbnail.height,
             mimeType = thumbnail.contentType.toString(),
-            size = thumbnail.file.size
+            size = thumbnail.file.size.toLong()
         )
     }
 
@@ -279,7 +279,7 @@ class MediaServiceImpl(
             width = thumbnail.width,
             height = thumbnail.height,
             mimeType = thumbnail.contentType.toString(),
-            size = thumbnail.file.size
+            size = thumbnail.file.size.toLong()
         )
     }
