+++ trixnity-core/src/commonMain/kotlin/net/folivo/trixnity/core/model/events/m/room/ImageInfo.kt
@@ -12,7 +12,7 @@ data class ImageInfo(
     @SerialName("mimetype")
     override val mimeType: String? = null,
     @SerialName("size")
-    override val size: Int? = null,
+    override val size: Long? = null,
     @SerialName("thumbnail_url")
     val thumbnailUrl: String? = null,
     @SerialName("thumbnail_file")
